Refs #4. "mcvine bash_aliases" -> "mcvine bash aliase". added "mcvine bash complete"
